@@ -1,6 +1,6 @@
 {
   "name": "react-storybook-addon-chapters",
-  "version": "1.1.0",
+  "version": "1.1.3",
   "description": "React Storybook Chapters addon allows showcasing of multiple components within a story by breaking it down into smaller categories (chapters) and subcategories (sections) for more organizational goodness.",
   "repository": {
     "type": "git",
@@ -20,7 +20,7 @@
     "publish-storybook": "bash .scripts/publish_storybook.sh"
   },
   "devDependencies": {
-    "@kadira/storybook": "^2.20.1",
+    "@storybook/addon-info": "^3.2.12",
     "babel-cli": "^6.5.0",
     "babel-core": "^6.5.0",
     "babel-eslint": "^7.0.0",
@@ -29,6 +29,7 @@
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-2": "^6.5.0",
+    "create-react-class": "^15.6.0",
     "chai": "^3.5.0",
     "enzyme": "^2.2.0",
     "eslint": "^3.6.1",
@@ -41,9 +42,9 @@
     "jsdom": "^9.5.0",
     "mocha": "^3.0.2",
     "prop-types": "^15.5.10",
-    "react": "^15.3.2",
-    "react-addons-test-utils": "^15.3.1",
-    "react-dom": "^0.14.7 || ^15.0.0",
+    "react": "^16.0.0",
+    "react-test-renderer": "^16.0.0-0",
+    "react-dom": "^0.14.7 || ^16.0.0",
     "sinon": "^1.17.6"
   },
   "peerDependencies": {
@@ -51,9 +52,11 @@
     "react": "^0.14.7 || ^15.0.0"
   },
   "dependencies": {
-    "@kadira/react-storybook-addon-info": "^3.3.0",
+    "@storybook/addon-info": "^3.2.12",
+    "@storybook/components": "^3.2.12",
+    "@storybook/react": "^3.2.12",
     "babel-runtime": "^6.5.0",
-    "markdown-to-react-components": "^0.2.1",
+    "markdown-to-react-components": "https://github.com/besrabasant/markdown-to-react-components.git#react-15.6.1",
     "react-addons-create-fragment": "^15.3.2"
   },
   "main": "dist/index.js",
