@@ -1,21 +1,22 @@
-import React from "react"
-import Button from "../Button"
-import {storiesOf} from "@kadira/storybook"
-import {SectionDecorator, StoryDecorator, sectionStyles, Theme} from "../../src/"
-import {Pre} from "@kadira/react-storybook-addon-info/dist/components/markdown"
-import "./styles.css"
+/*
+import { Pre } from '@kadira/react-storybook-addon-info/dist/components/markdown';
+import { storiesOf } from '@storybook/react';
+import React from 'react';
+import Button from '../Button';
+import { SectionDecorator, StoryDecorator, sectionStyles, Theme } from '../../src/';
+import './styles.css';
 
 // Override styles
 sectionStyles.header.marginBottom = 10;
 sectionStyles.buttonContainer.marginTop = 10;
 
 // Access colors from Theme
-const headerStyles = {color: Theme.grayDark};
+const headerStyles = { color: Theme.grayDark };
 
 // Custom story subtitle view
 StoryDecorator.subtitle = function (subtitle) {
   return (
-    <p className="story-subtitle" style={{color: Theme.grayDark}}>
+    <p className="story-subtitle" style={{ color: Theme.grayDark }}>
       {subtitle}
     </p>
   );
@@ -26,7 +27,7 @@ SectionDecorator.sourceCode = function (sourceCode) {
   return (
     <div className="source-box">
       <h4 style={headerStyles}>Source</h4>
-      <div style={{borderColor: Theme.gray, backgroundColor: Theme.grayLighter}}>
+      <div style={{ borderColor: Theme.gray, backgroundColor: Theme.grayLighter }}>
         <Pre>{sourceCode}</Pre>
       </div>
     </div>
@@ -38,32 +39,36 @@ SectionDecorator.component = function (component) {
   return (
     <div className="example-box">
       <h4 style={headerStyles}>Example</h4>
-      <div style={{borderColor: Theme.gray}}>{component}</div>
+      <div style={{ borderColor: Theme.gray }}>{component}</div>
     </div>
   );
 };
 
-storiesOf('Custom decorators')
+ storiesOf('Custom decorators')
   .addWithChapters(
-    'Story With Chapters',
-    {
-      subtitle: "And with customized layout",
-      chapters: [
-        {
-          sections: [
-            {
-              title: 'Default settings',
-              sectionFn: () => (<Button label="My Button" onClick={() => {
+  'Story With Chapters',
+  {
+    subtitle: 'And with customized layout',
+    chapters: [
+      {
+        sections: [
+          {
+            title: 'Default settings',
+            sectionFn: () => (<Button
+              label="My Button" onClick={() => {
                 alert('Hello World!');
-              }}/>),
-            },
-            {
-              title: 'Disabled',
-              sectionFn: () => (<Button label="My Disabled Button" disabled onClick={() => {
-              }}/>),
-            },
-          ],
-        },
-      ]
-    }
-  );
+              }}
+            />),
+          },
+          {
+            title: 'Disabled',
+            sectionFn: () => (<Button
+              label="My Disabled Button" disabled onClick={() => {
+              }}
+            />),
+          },
+        ],
+      },
+    ],
+  }
+  );*/
