@@ -1,10 +1,35 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import Node from '@kadira/react-storybook-addon-info/dist/components/Node';
-import { Pre } from '@kadira/react-storybook-addon-info/dist/components/markdown';
+import Node from '@storybook/addon-info/dist/components/Node';
+import { Pre } from '@storybook/addon-info/dist/components/markdown';
 import PropTable from './PropTable';
 import renderInfoContent from '../utils/info-content';
 import theme from '../theme';
+
+const propTypes = {
+  title: PropTypes.string,
+  subtitle: PropTypes.string,
+  info: PropTypes.string,
+  showSource: PropTypes.bool,
+  showPropTables: PropTypes.bool,
+  propTables: PropTypes.arrayOf(PropTypes.func),
+  children: PropTypes.oneOfType([
+    PropTypes.object,
+    PropTypes.array,
+  ]),
+  addonInfo: PropTypes.object,
+};
+
+const defaultProps = {
+  context: {},
+  title: '',
+  subtitle: '',
+  info: '',
+  showSource: true,
+  allowSourceToggling: true,
+  showPropTables: false,
+  allowPropTablesToggling: true,
+};
 
 export const sectionButtonStyles = {
   backgroundColor: 'transparent',
@@ -15,6 +40,7 @@ export const sectionButtonStyles = {
   float: 'right',
   marginLeft: 5,
   padding: '5px 10px',
+
 };
 
 export const sectionStyles = {
@@ -60,6 +86,91 @@ export const sectionStyles = {
   },
 };
 
+export class SectionDecorator {
+  static main(header, component, additional) {
+    return (
+      <div style={sectionStyles.container}>
+        {header}
+        {component}
+        {additional}
+      </div>
+    );
+  }
+
+  static header(header) {
+    return (
+      <div style={sectionStyles.header}>
+        <div>{header}</div>
+      </div>
+    );
+  }
+
+  static title(title) {
+    return (
+      <h3 style={sectionStyles.title}>{title}</h3>
+    );
+  }
+
+  static subtitle(subtitle) {
+    return (
+      <p style={sectionStyles.subtitle}>{subtitle}</p>
+    );
+  }
+
+  static component(component) {
+    return (
+      <div style={sectionStyles.componentContainer}>
+        {component}
+      </div>
+    );
+  }
+
+  static additional(additional) {
+    return (
+      <div>
+        <div>{additional}</div>
+      </div>
+    );
+  }
+
+  static sourceCode(sourceCode) {
+    return (
+      <div style={sectionStyles.subsection}>
+        <h4 style={sectionStyles.subsectionTitle}>Source</h4>
+        <Pre>
+          {sourceCode}
+        </Pre>
+      </div>
+    );
+  }
+
+  static propTables(propTables) {
+    return (
+      <div style={sectionStyles.subsection}>
+        <h4 style={sectionStyles.subsectionTitle}>PropTypes</h4>
+        {propTables}
+      </div>
+    );
+  }
+
+  static buttons(buttons) {
+    return (
+      <div style={sectionStyles.buttonContainer}>{buttons}</div>
+    );
+  }
+
+  static info(infoContent) {
+    return (
+      <div style={sectionStyles.subsection}>
+        <div style={sectionStyles.info}>
+          {infoContent}
+        </div>
+      </div>
+    );
+  }
+}
+
+
 export default class Section extends Component {
   constructor(props) {
     super(props);
@@ -70,8 +181,10 @@ export default class Section extends Component {
   }
 
   renderSourceCode() {
+    const addonInfo = this.props.addonInfo;
+
     const sourceCode = React.Children.map(this.props.children, (root, idx) => (
-      <Node key={idx} depth={0} node={root} />
+      <Node key={idx} depth={0} node={root} {...addonInfo} {...this.props} />
     ));
 
     return SectionDecorator.sourceCode(sourceCode);
@@ -147,22 +260,26 @@ export default class Section extends Component {
 
     const buttons = [
       this.props.allowPropTablesToggling &&
-        <button key="allowPropTablesToggling" onClick={() => {
+      <button
+        key="allowPropTablesToggling" onClick={() => {
           this.setState({
             isPropsTableShown: !this.state.isPropsTableShown,
           });
-        }} style={this.state.isPropsTableShown ? sectionStyles['button-active'] : sectionStyles.button }>
-          {this.state.isPropsTableShown ? 'Hide' : 'Show'} Props Table
+        }} style={this.state.isPropsTableShown ? sectionStyles['button-active'] : sectionStyles.button}
+      >
+        {this.state.isPropsTableShown ? 'Hide' : 'Show'} Props Table
         </button>,
 
       this.props.allowSourceToggling &&
-        <button key="allowSourceToggling" onClick={() => {
+      <button
+        key="allowSourceToggling" onClick={() => {
           this.setState({
             isSourceShown: !this.state.isSourceShown,
           });
-        }} style={this.state.isSourceShown ? sectionStyles['button-active'] : sectionStyles.button }>
-          {this.state.isSourceShown ? 'Hide' : 'Show'} Source
-        </button>
+        }} style={this.state.isSourceShown ? sectionStyles['button-active'] : sectionStyles.button}
+      >
+        {this.state.isSourceShown ? 'Hide' : 'Show'} Source
+        </button>,
     ];
 
     const additional = (
@@ -182,111 +299,5 @@ export default class Section extends Component {
   }
 }
 
-Section.displayName = 'Section';
-Section.propTypes = {
-  context: PropTypes.object,
-  title: PropTypes.string,
-  subtitle: PropTypes.string,
-  info: PropTypes.string,
-  showSource: PropTypes.bool,
-  showPropTables: PropTypes.bool,
-  propTables: PropTypes.arrayOf(PropTypes.func),
-  children: PropTypes.oneOfType([
-    PropTypes.object,
-    PropTypes.array,
-  ]),
-};
-Section.defaultProps = {
-  context: {},
-  title: '',
-  subtitle: '',
-  info: '',
-  showSource: true,
-  allowSourceToggling: true,
-  showPropTables: false,
-  allowPropTablesToggling: true,
-};
-
-export class SectionDecorator {
-  static main(header, component, additional) {
-    return (
-      <div style={sectionStyles.container}>
-        {header}
-        {component}
-        {additional}
-      </div>
-    );
-  };
-
-  static header(header) {
-    return (
-      <div style={sectionStyles.header}>
-        <div>{header}</div>
-      </div>
-    );
-  };
-
-  static title(title) {
-    return (
-      <h3 style={sectionStyles.title}>{title}</h3>
-    );
-  }
-
-  static subtitle(subtitle) {
-    return (
-      <p style={sectionStyles.subtitle}>{subtitle}</p>
-    );
-  }
-
-  static component(component) {
-    return (
-      <div style={sectionStyles.componentContainer}>
-        {component}
-      </div>
-    );
-  };
-
-  static additional(additional) {
-    return (
-      <div>
-        <div>{additional}</div>
-      </div>
-    );
-  };
-
-  static sourceCode(sourceCode) {
-    return (
-      <div style={sectionStyles.subsection}>
-        <h4 style={sectionStyles.subsectionTitle}>Source</h4>
-        <Pre>
-          {sourceCode}
-        </Pre>
-      </div>
-    );
-  };
-
-  static propTables(propTables) {
-    return (
-      <div style={sectionStyles.subsection}>
-        <h4 style={sectionStyles.subsectionTitle}>PropTypes</h4>
-        {propTables}
-      </div>
-    );
-  };
-
-  static buttons(buttons) {
-    return (
-      <div style={sectionStyles.buttonContainer}>{buttons}</div>
-    )
-  };
-
-  static info(infoContent) {
-    return (
-      <div style={sectionStyles.subsection}>
-        <div style={sectionStyles.info}>
-          {infoContent}
-        </div>
-      </div>
-    );
-  };
-}
+Section.propTypes = propTypes;
+Section.defaultProps = defaultProps;
