@@ -1,9 +1,26 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { baseFonts } from '@kadira/react-storybook-addon-info/dist/components/theme';
+import { baseFonts } from '@storybook/components';
 import Chapter from './Chapter';
 import renderInfoContent from '../utils/info-content';
 import theme from '../theme';
+
+const propTypes = {
+  context: PropTypes.object,
+  title: PropTypes.string,
+  subtitle: PropTypes.string,
+  info: PropTypes.string,
+  chapters: PropTypes.arrayOf(PropTypes.object),
+  addonInfo: PropTypes.object,
+};
+
+const defaultProps = {
+  context: {},
+  title: '',
+  subtitle: '',
+  info: '',
+  chapters: [],
+};
 
 export const storyStyles = {
   story: {
@@ -26,42 +43,6 @@ export const storyStyles = {
   info: theme.infoStyle,
 };
 
-export default class Story extends Component {
-  render() {
-    const { context, subtitle, title, info, chapters } = this.props;
-
-    const header = (
-      <div>
-        {title && StoryDecorator.title(title)}
-        {subtitle && StoryDecorator.subtitle(subtitle)}
-        {info && StoryDecorator.subtitle(renderInfoContent(info))}
-      </div>
-    );
-
-    const renderedChapters = chapters.map((chapter, index) => (
-      <Chapter key={index} context={context} {...chapter} />
-    ));
-
-    return StoryDecorator.main(header, renderedChapters);
-  }
-}
-
-Story.displayName = 'Story';
-Story.propTypes = {
-  context: PropTypes.object,
-  title: PropTypes.string,
-  subtitle: PropTypes.string,
-  info: PropTypes.string,
-  chapters: PropTypes.arrayOf(PropTypes.object),
-};
-Story.defaultProps = {
-  context: {},
-  title: '',
-  subtitle: '',
-  info: '',
-  chapters: [],
-};
-
 export class StoryDecorator {
   static title(title) {
     return (
@@ -71,7 +52,7 @@ export class StoryDecorator {
 
   static subtitle(subtitle) {
     return (
-      <p style={storyStyles.subtitle}>{subtitle}</p>
+      <span style={storyStyles.subtitle}>{subtitle}</span>
     );
   }
 
@@ -88,5 +69,28 @@ export class StoryDecorator {
         {chapters}
       </div>
     );
-  };
+  }
 }
+
+export default class Story extends Component {
+  render() {
+    const { context, subtitle, title, info, chapters, addonInfo } = this.props;
+
+    const header = (
+      <div>
+        {title && StoryDecorator.title(title)}
+        {subtitle && StoryDecorator.subtitle(subtitle)}
+        {info && StoryDecorator.subtitle(renderInfoContent(info))}
+      </div>
+    );
+
+    const renderedChapters = chapters.map((chapter, index) => (
+      <Chapter key={index} context={context} addonInfo={addonInfo} {...chapter} />
+    ));
+
+    return StoryDecorator.main(header, renderedChapters);
+  }
+}
+
+Story.propTypes = propTypes;
+Story.defaultProps = defaultProps;
